qi502: redo
