+++ Blasphemous.LostDreams/Items/QuestItems/QuestItemList.cs
@@ -3,4 +3,5 @@ namespace Blasphemous.LostDreams.Items.QuestItems;
 
 internal class QuestItemList(Config cfg) : ItemList<QuestItem>
 {
+    public QuestItem QI502 { get; } = new QuestItem(new QI502());
 }
